+++ packages/web/src/components/login/sign/TextBox.tsx
@@ -1,4 +1,7 @@
 const TextBox = () => {
+  /**
+   * Sign in/upフォームのテキストボックス
+   */
   return <>
     <div className="form-control w-full">
       <input
