+++ packages/web/src/styles/globals.scss
@@ -53,9 +53,16 @@ $breakpoint: (
     }
     // sign in/upフォームのコンテナ
     &-sign-container {
-      @apply mx-[20px] p-[20px] w-full bg-gray-200 rounded-[30px];
-      @include mq(tab) {
-        @apply max-w-[800px] p-[30px];
+      // 外側
+      &-outer {
+        @apply mt-4 w-[100vw] flex justify-center;
+      }
+      // 内側
+      &-inner {
+        @apply mx-[20px] p-[20px] w-full bg-gray-200 rounded-[30px];
+        @include mq(tab) {
+          @apply max-w-[800px] p-[30px];
+        }
       }
     }
     // 戻るボタン
